+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.0] - Unreleased
+### Added
+- `runString` function.
+
 ### Changed
 - Removed `=` character for printing tags.